refactor: separate fitness normalization from file decoding
