+++ docs/fitness-file-storage.md
@@ -52,7 +52,7 @@ Important columns include:
 - `isPrimary`, `importBatchId`, `importStatus`, `importError`
 - `totalDistanceMeters`, `totalDurationSeconds`, `elevationGainMeters`
 - `movingTimeSeconds` — time (seconds) the athlete was actually moving, with stops excluded. Kept separate from `totalDurationSeconds` (elapsed time) so average pace/speed is computed over moving time, matching how Strava reports it. Nullable: records parsed before this column existed, or files with no per-point data to derive it, fall back to elapsed time (backfill with `scripts/fitness/backfillFitnessMovingTime.ts`).
-- `activityType` — the sport, stored as a canonical **activity type** (`ride`, `gravel_ride`, `mountain_bike_ride`, `ebike_ride`, `virtual_ride`, `run`, `trail_run`, `walk`, `hike`, `swim`, `training`, `rowing`, `yoga`, `climbing`, `ski`, `skating`, `surfing`, `racket_sports`, `martial_arts`, `team_sports`, `golf`, `other`), not as the raw string the source file carried. Four vocabularies reach the parser — FIT `sport`/`sub_sport` (`cycling`, `gravel_cycling`), Garmin TCX `Sport` (`Biking`), Strava `sport_type` (`Ride`, `GravelRide`, written verbatim into the TCX built for an import) and free-form GPX `<trk><type>` text — so the same ride arrived spelled three ways and everything that GROUPS or FILTERS on the value counted it as three activities: the fitness overview breakdown, the calendar filter, and the per-type route-heatmap cache key. `toActivityData` in `parseFitnessFile.ts` is the single funnel every parser returns through, so normalizing there covers uploads and the Strava webhook alike. Any value that does not match a known sport key or canonical non-gear activity defaults to `other`. Nullable. History imported before this rule is swept by `scripts/fitness/normalizeFitnessActivityTypes.ts`.
+- `activityType` — the sport, stored as a canonical **activity type** (`ride`, `gravel_ride`, `mountain_bike_ride`, `ebike_ride`, `virtual_ride`, `run`, `trail_run`, `walk`, `hike`, `swim`, `training`, `rowing`, `yoga`, `climbing`, `ski`, `skating`, `surfing`, `racket_sports`, `martial_arts`, `team_sports`, `golf`, `other`), not as the raw string the source file carried. Four vocabularies reach the parser — FIT `sport`/`sub_sport` (`cycling`, `gravel_cycling`), Garmin TCX `Sport` (`Biking`), Strava `sport_type` (`Ride`, `GravelRide`, written verbatim into the TCX built for an import) and free-form GPX `<trk><type>` text — so the same ride arrived spelled three ways and everything that GROUPS or FILTERS on the value counted it as three activities: the fitness overview breakdown, the calendar filter, and the per-type route-heatmap cache key. `toActivityData` in `activityData.ts` (`parseFitnessFile.ts`) is the single funnel every parser returns through, so normalizing there covers uploads and the Strava webhook alike. Any value that does not match a known sport key or canonical non-gear activity defaults to `other`. Nullable. History imported before this rule is swept by `scripts/fitness/normalizeFitnessActivityTypes.ts`.
 - `activityStartTime`
 - `hasMapData`, `mapImagePath`
 - `mapError` — why the route map is missing, when one was expected. Deliberately separate from `processingStatus`/`importError`: those mean "this activity is not usable" and gate the status detail dashboard, the post's stat grid, the fitness overview, the profile's Fitness tab and every stats/heatmap rollup, so reusing them for a missing image would hide hundreds of good imports during a tile-server or storage outage. Nullable, and null is the norm — no GPS data, a privacy location leaving no exposed route, and a map that rendered fine all leave it null. Cleared on every successful (re)processing run. The post surfaces it to its **owner only**, and as a closed vocabulary rather than the reason — `fitness.mapFailure` is `missing` (none could be produced) or `stale` (the previous one is still shown, because the run that should have replaced it failed), so each surface's copy matches what the post actually shows. The status payload is readable by every viewer, while the reason is a raw error string that can name internal infrastructure, so it is rendered only on the owner's own fitness files page.
